# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cервер на Golang с OpenAPI документацией.",
        "title": "Shop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Maksim",
            "url": "https://github.com/hhullen",
            "email": "hhullen@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/supplier": {
            "get": {
                "description": "Возвращает поставщика",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supplier"
                ],
                "summary": "Возвращает поставщика",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"609ccf6f-7fb4-44bd-aa77-bc9e0e7572b4\"",
                        "description": "uid",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "true",
                        "description": "avoid_cache",
                        "name": "avoid_cache",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetSupplierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetSupplierResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetSupplierResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавление поставщика. Если поставщик существует, вернется uid этого поставщика.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supplier"
                ],
                "summary": "Добавление поставщика",
                "parameters": [
                    {
                        "description": "Информация о поставщике",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.AddSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.AddSupplierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.AddSupplierResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.AddSupplierResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаление поставщика",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supplier"
                ],
                "summary": "Удаление поставщика",
                "parameters": [
                    {
                        "description": "uid",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.DeleteSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.DeleteSupplierResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.DeleteSupplierResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.DeleteSupplierResponse"
                        }
                    }
                }
            }
        },
        "/supplier/address": {
            "patch": {
                "description": "Обновление адреса поставщика",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supplier"
                ],
                "summary": "Обновление адреса поставщика",
                "parameters": [
                    {
                        "description": "uid и адрес",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.UpdateSupplierAddressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.UpdateSupplierAddressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.UpdateSupplierAddressResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.UpdateSupplierAddressResponse"
                        }
                    }
                }
            }
        },
        "/suppliers": {
            "get": {
                "description": "Возвращает поставщиков",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supplier"
                ],
                "summary": "Возвращает поставщиков",
                "parameters": [
                    {
                        "type": "string",
                        "example": "0",
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "10",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "true",
                        "description": "avoid_cache",
                        "name": "avoid_cache",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetSuppliersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetSuppliersResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetSuppliersResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "datastruct.AddSupplierRequest": {
            "type": "object",
            "required": [
                "address",
                "name",
                "phone_number"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/datastruct.Address"
                },
                "avoid_cache": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Vasilisa\u0026Drozzhi .ltd"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79336579933 RU"
                },
                "uid": {
                    "type": "string",
                    "example": "609ccf6f-7fb4-44bd-aa77-bc9e0e7572b4"
                }
            }
        },
        "datastruct.AddSupplierResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "datastruct.Address": {
            "type": "object",
            "required": [
                "city",
                "country",
                "street"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Seattle"
                },
                "country": {
                    "type": "string",
                    "example": "USA"
                },
                "street": {
                    "type": "string",
                    "example": "12th Ave E"
                }
            }
        },
...
        "datastruct.DeleteSupplierRequest": {
            "type": "object",
            "required": [
                "uid"
            ],
            "properties": {
                "avoid_cache": {
                    "type": "boolean",
                    "example": true
                },
                "uid": {
                    "type": "string",
                    "example": "609ccf6f-7fb4-44bd-aa77-bc9e0e7572b4"
                }
            }
        },
        "datastruct.DeleteSupplierResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
...
        "datastruct.GetSupplierResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                },
                "supplier": {
                    "$ref": "#/definitions/datastruct.Supplier"
                }
            }
        },
        "datastruct.GetSuppliersResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "suppliers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/datastruct.Supplier"
                    }
                }
            }
        },
...
        "datastruct.Supplier": {
            "type": "object",
            "required": [
                "address",
                "name",
                "phone_number"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/datastruct.Address"
                },
                "name": {
                    "type": "string",
                    "example": "Vasilisa\u0026Drozzhi .ltd"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79336579933 RU"
                },
                "uid": {
                    "type": "string",
                    "example": "609ccf6f-7fb4-44bd-aa77-bc9e0e7572b4"
                }
            }
        },
...
        "datastruct.UpdateSupplierAddressRequest": {
            "type": "object",
            "required": [
                "address",
                "uid"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/datastruct.Address"
                },
                "avoid_cache": {
                    "type": "boolean",
                    "example": true
                },
                "uid": {
                    "type": "string",
                    "example": "609ccf6f-7fb4-44bd-aa77-bc9e0e7572b4"
                }
            }
        },
        "datastruct.UpdateSupplierAddressResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        }
    }
}